+++ .gitignore
@@ -29,6 +29,7 @@ build/Release
 
 # Dependency directories
 .idea
+dist
 node_modules
 jspm_packages
