+++ src/components/authors/authorPage.js
@@ -5,8 +5,9 @@
 
 var React = require('react');
 var AuthorApi = require('../../api/authorApi');
+var AuthorList = require('./authorList');
 
-var Authors = React.createClass({
+var AuthorPage = React.createClass({
     // set initial state
     getInitialState: function() {
         return {
@@ -15,39 +16,21 @@ var Authors = React.createClass({
     },
     // get Data for state, here Mock
     // ca be ajax, or other calls
-    componentWillMount: function() {
-        this.setState({authors: AuthorApi.getAllAuthors()});
+    componentDidMount: function() {
+        if (this.isMounted()) {
+            this.setState({authors: AuthorApi.getAllAuthors()});
+        }
     },
 
     render: function() {
-        // for each author draw this snippet
-        var createAuthorRow = function(author) {
-            return (
-                <tr key={author.id}>
-                    <td><a href={"/#authors/" + author.id}>{author.id}</a></td>
-                    <td>{author.firstName} {author.lastName}</td>
-                </tr>
-            );
-        };
-
         return(
             // div is the wrapper (top-level-component)
             <div>
                 <h1>Authors</h1>
-
-                <table className="table">
-                    <thead>
-                        <th>ID</th>
-                        <th>Name</th>
-                    </thead>
-                    <tbody>
-                        // iterate through data in state
-                        {this.state.authors.map(createAuthorRow, this)}
-                    </tbody>
-                </table>
+                <AuthorList authors={this.state.authors} />
             </div>
         );
     }
 });
 
-module.exports = Authors;
+module.exports = AuthorPage;
